Added feature for all categories case.

+++ yolo2mot.py
@@ -4,6 +4,7 @@
 # load libraries
 import glob
 import numpy as np
+import pandas as pd
 from tqdm import tqdm
 from skimage import io
 import cv2
@@ -28,7 +29,7 @@ def parse_args():
     parser = argparse.ArgumentParser(description='Running experiment')
     parser.add_argument("--class_cat", help="Select the class category", type=str)
     parser.add_argument("--seq_name", help="Select the sequence name", type=str)
-    parser.add_argument("--class_id", help="Select the class ID", type=int)
+    parser.add_argument("--class_id", help="Select the class ID", type=str) # str is default type
     parser.add_argument("--source_path", help="source path to the images", type=str, default="/local-scratch/share_dataset/labled_hevc_sequences")
     parser.add_argument("--test", help="test for public dataset",  action='store_true')
     args = parser.parse_args()
@@ -72,27 +73,38 @@ if __name__ == '__main__':
 
     try:
         if class_cat_filter == "all":
+            
             # ----------------------------------------------------------------
             # Generate yolo v3 output file in MOT format for all object categories
             # ----------------------------------------------------------------
             print("Converting yolo v3 output file in MOT format ...")
-            alltxt = glob.glob(f"{input_path}")    
+            alltxt = natural_sort(glob.glob(f"{input_path}"))
             if os.path.exists(output_path):
                 os.remove(output_path)
-            frame = 1
             with open(output_path, 'ab') as output_file:
                 for txt in tqdm(alltxt):
+                    if test_flag:
+                        frame = int(txt.split("/")[-1].split(".")[0]) 
+                    else:
+                        frame = int(txt.split("_")[-1].split(".")[0]) + 1      
                     #print(f"for frame {frame} ...")
-                    class_id, x, y, w, h = np.genfromtxt(txt, delimiter=' ', unpack=True, encoding='utf_8_sig')
+                    class_id, x, y, w, h = np.genfromtxt(txt, unpack=True, encoding='utf_8_sig')
 
-                    if not isinstance(class_id, np.ndarray):
-                        class_id = np.array([class_id])
-                    frame_id = np.repeat(frame, len(class_id))
-                    object_id = np.repeat(-1, len(class_id)) # no tracking, so assign -1
-                    conf_arr = np.repeat(1, len(class_id))
-                    pos_x_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
-                    pos_y_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
-                    pos_z_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                    # Either numpy array or single value if you read one row
+                    if isinstance(class_id, np.ndarray):
+                        frame_id = np.repeat(frame, len(class_id))
+                        object_id = np.repeat(-1, len(class_id)) # no tracking, so assign -1
+                        conf_arr = np.repeat(1, len(class_id))
+                        pos_x_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                        pos_y_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                        pos_z_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                    else:
+                        frame_id = frame
+                        object_id = -1 # no tracking, so assign -1
+                        conf_arr = 1
+                        pos_x_3d = -1 # not 3d, so assign -1
+                        pos_y_3d = -1 # not 3d, so assign -1
+                        pos_z_3d = -1 # not 3d, so assign -1
 
                     # conversion of bbox
                     x = normalize(x, 0, 1, 0, img_w)
@@ -105,40 +117,44 @@ if __name__ == '__main__':
                     y1 = y - h / 2
                     # x2 = x + w / 2
                     # y2 = y + h / 2 
-
-
-                    # following line doesn't work for a single row when doing np.savetxt
-                    # data = np.array(np.transpose(\
-                    #     [frame_arr, object_id, x1, y1, x2, y2, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d]))
-
+                      
                     # save into det file
                     np.savetxt(output_file, np.column_stack(\
                         (frame_id, object_id, x1, y1, w, h, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d)\
                             ), delimiter=',', fmt="%d,%d,%s,%s,%s,%s,%d,%d,%d,%d")
-                    frame += 1
             
 
             # ----------------------------------------------------------------
             # Generate GT file in MOT format for all object categories
             # ----------------------------------------------------------------
             print("Converting GT file in MOT format ...")
-            alltxt_gt = glob.glob(f"{input_path_gt}")
+            if test_flag:
+                print("We don't convert public GT file!")
+                exit()
+            alltxt_gt = natural_sort(glob.glob(f"{input_path_gt}"))
             if os.path.exists(output_path_gt):
                 os.remove(output_path_gt)
-            frame = 1
+            classmap = dict()
+            obj_count = 0
             with open(output_path_gt, 'ab') as output_file:
                 for txt in tqdm(alltxt_gt):
-                    #print(f"for frame {frame} ...")
-                    class_id, object_id, x, y, w, h = np.genfromtxt(txt, delimiter=' ', unpack=True, encoding='utf_8_sig')
+                    frame = int(txt.split("_")[-1].split(".")[0]) + 1
+                    class_id, object_id, x, y, w, h = np.genfromtxt(txt, unpack=True, encoding='utf_8_sig')
+                    # no delmiter=' ' since the default is any whitespace
 
-
-                    if not isinstance(class_id, np.ndarray):
-                        class_id = np.array([class_id])
-                    frame_id = np.repeat(frame, len(class_id))
-                    conf_arr = np.repeat(1, len(class_id))
-                    pos_x_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
-                    pos_y_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
-                    pos_z_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                    # Either numpy array or single value if you read one row
+                    if isinstance(class_id, np.ndarray):
+                        frame_id = np.repeat(frame, len(class_id))
+                        conf_arr = np.repeat(1, len(class_id))
+                        pos_x_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                        pos_y_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                        pos_z_3d = np.repeat(-1, len(class_id)) # not 3d, so assign -1
+                    else:
+                        frame_id = frame
+                        conf_arr = 1
+                        pos_x_3d = -1 # not 3d, so assign -1
+                        pos_y_3d = -1 # not 3d, so assign -1
+                        pos_z_3d = -1 # not 3d, so assign -1
 
                     # conversion of bbox
                     x = normalize(x, 0, 1, 0, img_w)
@@ -153,11 +169,42 @@ if __name__ == '__main__':
                     # y2 = y + h / 2 
 
                     # obj_id conversion for GT file, need some complex mapping algorithm
-                    # put into 2d array
-                    #data = np.transpose(np.array([frame_arr, class_id, object_id, x1, y1, w, h, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d]))
-                    #data = npdata[np.argsort(data[:, 1])] # sorting by class_id
-                    #data[:,2] = np.arange(1,len(data[:,2])+1,1) # assign id from 1
-                    #frame_arr, class_id, object_id, x1, y1, w, h, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d = data.T
+                    data = np.array([frame_id, class_id, object_id, x1, y1, w, h, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d]).T
+                    if data.ndim > 1:
+                        # go through row by row
+                        for i in range(len(data[:,1])):
+                            # add new class id in classmap if it doesn't exist
+                            if data[i,1] not in classmap:
+                                classmap[data[i,1]] = dict()
+                            old_obj_id = data[i,2]
+
+                            # if old object id is in classmap[class_id] dictionary
+                            # new object found!
+                            if old_obj_id not in classmap[data[i,1]]:
+                                obj_count += 1
+                                new_obj_id = obj_count
+                                data[i,2] = new_obj_id
+                                classmap[data[i,1]][old_obj_id] = new_obj_id
+                            # previously found object again
+                            else:
+                                new_obj_id = classmap[data[i,1]][old_obj_id]
+                                data[i,2] = new_obj_id
+                    
+                    # in case of only 1 row of numpy array
+                    else:
+                        if data[1] not in classmap:
+                            classmap[data[1]] = dict()
+                        old_obj_id = data[2]
+                        if old_obj_id not in classmap[data[1]]:
+                            obj_count += 1
+                            new_obj_id = obj_count
+                            data[2] = new_obj_id
+                            classmap[data[1]][old_obj_id] = new_obj_id
+                        else:
+                            new_obj_id = classmap[data[1]][old_obj_id]
+                            data[2] = new_obj_id
+
+                    frame_id, class_id, object_id, x1, y1, w, h, conf_arr, pos_x_3d, pos_y_3d, pos_z_3d = data.T
 
                     # save into det file
                     np.savetxt(output_file, np.column_stack(\
@@ -170,6 +217,7 @@ if __name__ == '__main__':
             # Generate yolo v3 output file in MOT format
             # ----------------------------------------------------------------
             print("Converting yolo v3 output file in MOT format ...")
+            class_cat_filter = int(class_cat_filter) # we have to use integer to filter numpy 2d array
             alltxt = natural_sort(glob.glob(f"{input_path}"))
             if os.path.exists(output_path):
                 os.remove(output_path)
